backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
pypdf
google-generativeai>=0.8.0
